% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iRafNet.R
\name{iRafNet_permutation}
\alias{iRafNet_permutation}
\title{Integrative random forest for co-expression network inference}
\usage{
iRafNet_permutation(X, W, ntree=NULL, mtry=NULL, genes.name,
                  seed, to.store=NULL)
}
\arguments{
\item{X}{\code{(n x p)} Matrix containing expression levels for \code{n} samples and \code{p} genes.}

\item{W}{\code{(p x p)} Symmatrix matrix containing iRafNet sampling scores. Element \code{(i,j)} contains score for interaction \code{(i - j)}. Scores must be non-negative. Larger value of sampling score corresponds to higher likelihood of gene \code{i} interacting with gene \code{j}. Columns and rows of \code{W} must be in the same order as the columns of \code{X}. Sampling scores \code{W} are computed considering one prior data such as protein-protein interactions.}

\item{ntree}{Numeric value: number of trees. If omitted, \code{ntree} is set to \code{1000}.}

\item{mtry}{Numeric value: number of predictors to be sampled at each node. If omitted, \code{mtry} is set to the square root of the number of variables.}

\item{genes.name}{Vector containing genes name. The order needs to match the rows of \code{x_j}.}

\item{seed}{Integer: Seed of permutation.}

\item{to.store}{Optional Integer. Total number of importance scores to be stored. When omitted, all importance scores will be stored. Note that to compute FDR we do not need all \code{(p-p) x p / 2} importance scores where \code{p} is the total number of proteins/genes, a sufficiently large number would work. This number is usually chosen based on the number of nodes. Suggested value is \code{p x 20}. }
}
\value{
Importance score for each regulatory relationship. The first column contains gene name of regulators, the second column contains gene name of targets, and third column contains corresponding importance scores.
}
\description{
This function fits iRafNet, a flexible unified integrative algorithm that allows information from prior data, such as protein-protein interactions and gene knock-down, to be jointly considered for gene regulatory network inference. This function takes as input only one set of sampling scores, computed considering one prior data such as protein-protein interactions or gene expression from knock-out experiments. Note that some of the functions utilized are a modified version of functions contained in the R package randomForest (A. Liaw and M. Wiener, 2002). }
\examples{

  # --- Generate data sets
  n<-20                  # sample size
  p<-5                   # number of genes
  genes.name<-paste("G",seq(1,p),sep="")   # genes name
  data<-matrix(rnorm(p*n),p,n)      # generate expression matrix
  W<-abs(matrix(rnorm(p*p),p,p))    # generate weights for interactions
 
  # --- Run iRafNet and obtain importance score of interactions
  out<-iRafNet_permutation(X=list(data),W=W,genes.name=genes.name,seed=1)

}
\references{
Petralia, F., Wang, P., Yang, J., Tu, Z. (2015) Integrative random forest for gene regulatory network inference, \emph{Bioinformatics}, \bold{31}, i197-i205.

A. Liaw and M. Wiener (2002). Classification and Regression by randomForest. \emph{R News} \bold{2}, 18--22.
}
