% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JRF.R
\name{ptmJRF_permutation}
\alias{ptmJRF_permutation}
\title{Joint Random Forest for the simultaneous estimation of interaction networks based on gene expression data, global proteomics data and post translational modification (PTM) data.}
\usage{
ptmJRF_permutation(X, ntree=NULL, mtry=NULL, genes.name,
                 ptm.name, seed, to.store=NULL)
}
\arguments{
\item{X}{List object containing expression data for each class, \code{X=list(x_1,x_2, ... )} where \code{x_1} is a \code{(F x n_j)} matrix with rows corresponding to post translational modification sites and columns to samples, while  \code{x_j} for  \code{j > 1} is a \code{(p x n_j)} matrix with rows corresponding to proteins and columns to samples. For \code{x_2, x_3, ...} Rows need to be the same corresponding to the same proteins, while samples can vary. Missing values are not allowed. Rows of object \code{x_1} does not need to be ordered in a specific way.
}

\item{ntree}{Numeric value: number of trees. }

\item{mtry}{Numeric value: number of predictors to be sampled at each node. }

\item{genes.name}{Vector containing genes name. The order needs to match the rows of \code{x_j}.}

\item{ptm.name}{List of post translational modification variables in protein domain. This list must be ordered as rows of \code{X[[1]]}. }

\item{seed}{Integer. Permutation seed.}

\item{to.store}{Optional Integer. Total number of importance scores to be stored. When omitted, all importance scores will be stored. Note that to compute FDR we do not need all \code{(p-p) x p / 2} importance scores where \code{p} is the total number of proteins/genes, a sufficiently large number would work. This number is usually chosen based on the number of nodes. Suggested value is \code{p x 20}. }

}
\value{
A matrix with \code{I} rows and \code{C + 2} columns where \code{I} is the total number of gene-gene interactions and \code{C} is the number of classes. The first two columns contain gene names for each interaction while the remaining columns contain importance scores for different classes. 

\item{to.store}{Optional Integer. Total number of importance scores to be stored. When omitted, all importance scores will be stored. Note that to compute FDR we do not need all \code{(p-p) x p / 2} importance scores where \code{p} is the total number of proteins/genes, a sufficiently large number would work. This number is usually chosen based on the number of nodes. Suggested value is \code{p x 20}. }
}
\description{
Algorithm for the simultaneous estimation of multiple related networks. Some of the functions utilized are a modified version of functions contained in the R package randomForest (A. Liaw and M. Wiener, 2002). 
}
\examples{

 # --- Generate data sets
 nclasses=2               # number of data sets / classes
 n1<-n2<-20               # sample size for each data sets
 p<-5                   # number of variables (genes)
 genes.name<-paste("G",seq(1,p),sep="")   # genes name
 ptm.name<-c("G1","G2","G3","G3","G4","G5","G1")   # ptm name
 p.ptm<-length(ptm.name)
 
 data1<-matrix(rnorm(p.ptm*n2),p.ptm,n1)       # generate PTM data
 data2<-matrix(rnorm(p*n1),p,n1)       # generate global proteomics
 
 # --- Run JRF and obtain importance score of interactions
  out<-ptmJRF(X=list(data1,data2),genes.name=genes.name,
          ptm.name=ptm.name)

}
\references{
Petralia, F., Song, W.M., Tu, Z. and Wang, P. (2016). New method for joint network analysis reveals common and different coexpression patterns among genes and proteins in breast cancer. \emph{Journal of proteome research}, \bold{15}(3), pp.743-754.

A. Liaw and M. Wiener (2002). Classification and Regression by randomForest. \emph{R News} \bold{2}, 18--22.
}
